requirements/dev.txt
-r base.txt

django-debug-toolbar
requests